plat/arm: mhu: Move to drivers/ folder
