+++ drivers/arm/css/mhu/css_mhu_doorbell.c
@@ -7,9 +7,8 @@
 #include <platform_def.h>
 
 #include <arch_helpers.h>
+#include <drivers/arm/css/css_mhu_doorbell.h>
 #include <drivers/arm/css/scmi.h>
-
-#include "css_mhu_doorbell.h"
 
 void mhu_ring_doorbell(struct scmi_channel_plat_info *plat_info)
 {
