+++ plat/arm/css/common/css_common.mk
@@ -26,16 +26,16 @@ BL31_SOURCES		+=	plat/arm/css/common/css_pm.c			\
 				plat/arm/css/common/css_topology.c
 
 ifeq (${CSS_USE_SCMI_SDS_DRIVER},0)
-BL31_SOURCES		+=	plat/arm/css/drivers/scp/css_pm_scpi.c		\
-				plat/arm/css/drivers/mhu/css_mhu.c		\
+BL31_SOURCES		+=	drivers/arm/css/mhu/css_mhu.c			\
+				plat/arm/css/drivers/scp/css_pm_scpi.c		\
 				plat/arm/css/drivers/scpi/css_scpi.c
 else
-BL31_SOURCES		+=	drivers/arm/css/scmi/scmi_ap_core_proto.c	\
+BL31_SOURCES		+=	drivers/arm/css/mhu/css_mhu_doorbell.c		\
+				drivers/arm/css/scmi/scmi_ap_core_proto.c	\
 				drivers/arm/css/scmi/scmi_common.c		\
 				drivers/arm/css/scmi/scmi_pwr_dmn_proto.c	\
 				drivers/arm/css/scmi/scmi_sys_pwr_proto.c	\
-				plat/arm/css/drivers/scp/css_pm_scmi.c		\
-				plat/arm/css/drivers/mhu/css_mhu_doorbell.c
+				plat/arm/css/drivers/scp/css_pm_scmi.c
 endif
 
 # Process CSS_LOAD_SCP_IMAGES flag
@@ -55,12 +55,12 @@ ifeq (${CSS_LOAD_SCP_IMAGES},1)
     BL2_SOURCES		+=	plat/arm/css/drivers/scp/css_sds.c	\
 				plat/arm/css/drivers/sds/sds.c
   else
-    BL2U_SOURCES	+=	plat/arm/css/drivers/scp/css_bom_bootloader.c	\
-				plat/arm/css/drivers/mhu/css_mhu.c		\
+    BL2U_SOURCES	+=	drivers/arm/css/mhu/css_mhu.c			\
+				plat/arm/css/drivers/scp/css_bom_bootloader.c	\
 				plat/arm/css/drivers/scpi/css_scpi.c
 
-    BL2_SOURCES		+=	plat/arm/css/drivers/scp/css_bom_bootloader.c	\
-				plat/arm/css/drivers/mhu/css_mhu.c		\
+    BL2_SOURCES		+=	drivers/arm/css/mhu/css_mhu.c			\
+				plat/arm/css/drivers/scp/css_bom_bootloader.c	\
 				plat/arm/css/drivers/scpi/css_scpi.c
     # Enable option to detect whether the SCP ROM firmware in use predates version
     # 1.7.0 and therefore, is incompatible.
