+++ plat/arm/css/common/sp_min/css_sp_min.mk
@@ -9,13 +9,13 @@ BL32_SOURCES		+=	plat/arm/css/common/css_pm.c			\
 				plat/arm/css/common/css_topology.c
 
 ifeq (${CSS_USE_SCMI_SDS_DRIVER},0)
-BL32_SOURCES		+=	plat/arm/css/drivers/scp/css_pm_scpi.c		\
-				plat/arm/css/drivers/mhu/css_mhu.c		\
+BL32_SOURCES		+=	drivers/arm/css/mhu/css_mhu.c			\
+				plat/arm/css/drivers/scp/css_pm_scpi.c		\
 				plat/arm/css/drivers/scpi/css_scpi.c
 else
-BL32_SOURCES		+=	drivers/arm/css/scmi/scmi_common.c		\
+BL32_SOURCES		+=	drivers/arm/css/mhu/css_mhu_doorbell.c		\
+				drivers/arm/css/scmi/scmi_common.c		\
 				drivers/arm/css/scmi/scmi_pwr_dmn_proto.c	\
 				drivers/arm/css/scmi/scmi_sys_pwr_proto.c	\
-				plat/arm/css/drivers/scp/css_pm_scmi.c		\
-				plat/arm/css/drivers/mhu/css_mhu_doorbell.c
+				plat/arm/css/drivers/scp/css_pm_scmi.c
 endif
